// src/types/types.ts
export type Track = {
  contextTrack: {
    uri: string;
    uid: string;
    metadata: {
      artist_name: string;
      artist_uri: string;
      album_title: string;
      'collection.can_add': string;
      image_url: string;
      duration: string;
      album_artist_name: string;
      title: string;
      marked_for_download: string;
      'collection.can_ban': string;
      'actions.skipping_next_past_track': string;
      queued_by: string;
      image_xlarge_url: string;
      is_queued: string;
      image_large_url: string;
      album_uri: string;
      'collection.in_collection': string;
      has_lyrics: string;
      'actions.skipping_prev_past_track': string;
      track_player: string;
      image_small_url: string;
      'collection.is_banned': string;
    };
  };
};